+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:4.0
+// swift-tools-version:5.0
 import PackageDescription
 
 let package = Package(
@@ -10,8 +10,8 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/Alamofire/Alamofire.git", .upToNextMajor(from: "4.1.0")),
-        .package(url: "https://github.com/antitypical/Result.git", .upToNextMajor(from: "4.0.0")),
-        .package(url: "https://github.com/ReactiveCocoa/ReactiveSwift.git", .upToNextMajor(from: ("4.0.0"))),
+        .package(url: "https://github.com/antitypical/Result.git", .upToNextMajor(from: "4.1.0")),
+        .package(url: "https://github.com/ReactiveCocoa/ReactiveSwift.git", .upToNextMajor(from: ("5.0.0"))),
         .package(url: "https://github.com/ReactiveX/RxSwift.git", .upToNextMajor(from: "4.0.0"))
     ],
     targets: [
@@ -42,6 +42,5 @@ let package = Package(
                 "Tests",
                 "Sources/Supporting Files",
                 "Examples"])
-    ],
-    swiftLanguageVersions: [3, 4]
+    ]
 )