+++ Sources/Moya/Image.swift
@@ -1,7 +1,7 @@
-#if os(iOS) || os(watchOS) || os(tvOS)
+#if canImport(UIKit)
     import UIKit.UIImage
     public typealias ImageType = UIImage
-#elseif os(macOS)
+#elseif canImport(AppKit)
     import AppKit.NSImage
     public typealias ImageType = NSImage
 #endif